+++ src/EscrowFactory.sol
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.20;
+
 import {Escrow} from "./Escrow.sol";
 
 contract EscrowFactory {
@@ -7,7 +8,7 @@ contract EscrowFactory {
     Escrow[] public escrows;
     address public dao;
     address public owner;
-    mapping(address => uint) public balance;
+    mapping(address => uint256) public balance;
     address[] public tokens;
     mapping(address => address[]) public userCreatedEscrow;
     mapping(address => mapping(address => address[])) public userToUserEscrow;
@@ -19,27 +20,16 @@ contract EscrowFactory {
         address _tokenUsed;
     }
 
-    function onlyEscrowAllowedToken(
-        address _tokenUsed
-    ) private view returns (bool) {
-        for (uint index = 0; index < tokens.length; index++) {
+    function onlyEscrowAllowedToken(address _tokenUsed) private view returns (bool) {
+        for (uint256 index = 0; index < tokens.length; index++) {
             if (tokens[index] == _tokenUsed) return true;
         }
         return false;
     }
 
     function createEscrow(CreateEscrow memory _escrow) external {
-        require(
-            onlyEscrowAllowedToken(_escrow._tokenUsed),
-            "token not supported."
-        );
-        Escrow escrow = new Escrow(
-            _escrow._creator,
-            _escrow._otherParty,
-            _escrow._amount,
-            _escrow._tokenUsed,
-            dao
-        );
+        require(onlyEscrowAllowedToken(_escrow._tokenUsed), "token not supported.");
+        Escrow escrow = new Escrow(_escrow._creator, _escrow._otherParty, _escrow._amount, _escrow._tokenUsed, dao);
         escrows.push(escrow);
         allEscrowsAddresses.push(address(escrow));
         userCreatedEscrow[msg.sender].push(address(escrow));
