+++ src/interfaces/IEscrowFactory.sol
@@ -2,6 +2,6 @@
 pragma solidity ^0.8.20;
 
 interface IEscrowFactory {
-    function balance(address user) external returns (uint);
+    function balance(address user) external returns (uint256);
     // function
 }
